% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accProImageRestoration}
\alias{accProImageRestoration}
\title{image restoration}
\usage{
accProImageRestoration(M, lambda, delta, stepSize)
}
\arguments{
\item{M}{the grayscale matrix with some elements missing}

\item{lambda}{the shrinkage parameter}

\item{delta}{an upper bound of the second-order gradient of the 
first part of the optimization function}

\item{stepsize}{Acceleration parameter}
}
\value{
the reconstructed matrix
}
\description{
image restoration
}
