\name{imageRestoration-package}
\alias{imageRestoration-package}
\alias{imageRestoration}
\docType{package}
\title{
\packageTitle{imageRestoration}
}
\description{
\packageDescription{imageRestoration}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{imageRestoration}
\packageIndices{imageRestoration}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{imageRestoration}

Maintainer: \packageMaintainer{imageRestoration}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
