% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{softThresholdOperator}
\alias{softThresholdOperator}
\title{optimization problem: $\frac{1}{2}||Y-X||^2_F+\lambda||Y||_{S_1}$}
\usage{
softThresholdOperator(X, lambda)
}
\arguments{
\item{X}{the known matrix in the optimization problem}

\item{lambda}{the shrinkage parameter}
}
\value{
the optimal solution of the problem
}
\description{
optimization problem: $\frac{1}{2}||Y-X||^2_F+\lambda||Y||_{S_1}$
}
